refactor: extract change requests setting into separate component
Create ChangeRequestsApprovalsSetting component following the same
pattern as other settings (PreventFlagDefaultsSetting,
CaseSensitivitySetting, FeatureNameValidation).

Changes:
- New ChangeRequestsApprovalsSetting.tsx component
- AdditionalSettings simplified from 76 to 25 lines
- All settings now follow consistent pattern (receive project prop)

Benefits:
- Better consistency across all setting components
- Improved encapsulation (component owns its state)
- Easier to test in isolation
- Cleaner AdditionalSettings parent component
