+++ frontend/web/components/pages/project-settings/tabs/general-tab/sections/additional-settings/ChangeRequestsApprovalsSetting.tsx
@@ -1,0 +1,68 @@
+import React, { useCallback, useState } from 'react'
+import ChangeRequestsSetting from 'components/ChangeRequestsSetting'
+import Utils from 'common/utils/utils'
+import { Project } from 'common/types/responses'
+import { useUpdateProjectWithToast } from 'components/pages/project-settings/hooks'
+
+type ChangeRequestsApprovalsSettingProps = {
+  project: Project
+}
+
+export const ChangeRequestsApprovalsSetting = ({
+  project,
+}: ChangeRequestsApprovalsSettingProps) => {
+  const [updateProjectWithToast, { isLoading: isSaving }] =
+    useUpdateProjectWithToast()
+  const [minimumChangeRequestApprovals, setMinimumChangeRequestApprovals] =
+    useState(project.minimum_change_request_approvals)
+
+  const changeRequestsFeature = Utils.getFlagsmithHasFeature(
+    'segment_change_requests',
+  )
+
+  const saveChangeRequests = useCallback(async () => {
+    if (isSaving) return
+
+    await updateProjectWithToast(
+      {
+        minimum_change_request_approvals: minimumChangeRequestApprovals,
+        name: project.name,
+      },
+      project.id,
+      {
+        errorMessage: 'Failed to save project. Please try again.',
+        successMessage: 'Project Saved',
+      },
+    )
+  }, [
+    minimumChangeRequestApprovals,
+    project.name,
+    project.id,
+    isSaving,
+    updateProjectWithToast,
+  ])
+
+  const handleChangeRequestsToggle = useCallback(
+    (value: number | null) => {
+      setMinimumChangeRequestApprovals(value)
+      // Auto-save on toggle
+      saveChangeRequests()
+    },
+    [saveChangeRequests],
+  )
+
+  if (!changeRequestsFeature) {
+    return null
+  }
+
+  return (
+    <ChangeRequestsSetting
+      feature='4_EYES_PROJECT'
+      value={minimumChangeRequestApprovals}
+      onToggle={handleChangeRequestsToggle}
+      onSave={saveChangeRequests}
+      onChange={setMinimumChangeRequestApprovals}
+      isLoading={isSaving}
+    />
+  )
+}
